Remove "banned postcode" feature
This may be temporary!

Reverts https://github.com/alphagov/frontend/pull/1277 to see if this is still an issue. If it is, then I will reinstate with some instrumentation to permit inspection of the use of this feature.

We'll be able to see new instances of this occurring in the logs for locations-api. We short circuit the request before making API calls at present, hence the need for this change.