Give council tax search results page a more descriptive title in <head>
Make the title in the <head> more descriptive so that it doesn't duplicate the council tax search index page title. The new results page title is consistent with the title format on other pages where we add the chapter title after the page title: [page title]: [chapter title] - GOV.UK.

This issue with duplicated page titles was identified in a recent DAC audit of GOV.UK pages. Screen reader users had to explore the page to find out that the search results page was a different page and that something had changed on it.