Use a time tag to include the full date
Affects www.gov.uk/bank-holidays and www.gov.uk/when-do-the-clocks-change

The year and the month/day are spaced apart in the current implementation. The year is the caption to the table in which the month/day are listed. By using the `time` tag, we can[make this easier for machines to interpret](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/time).

The Google featured snippet for bank holidays sometimes shows the incorrect section of the page. It is hoped that we might improve this behaviour by providing a little assistance.

We control the content that gets displayed (in lib/data/*.json), so it should be fine to use html_safe to inject the tag into the table component.

The data is currently in yyyy/MM/dd format, but this gets converted to the desired yyyy-MM-dd format already. I've added a test to hopefully ensure that this remains the case.