Fix bug where location error was thrown when multiple authorities
This line was throwing a location error when there were multiple authorities returned. This check I've added is the same as the check on line 21 to check if the multiple authorities template should be rendered (that line also checks if postcode is provided, but we already check for that on line 88 here).

However, I'm not sure if my "fix" is the best way to handle this or if we should in fact investigate why the places_not_found method from the model doesn't recognise multiple authorities as a valid result. I've tried to superficially compare how this is handled for the other postcode finders like local transaction but those use the location_api model instead.