Fix bug that created a location error for multiple authorities
This line was creating a location error when the search result contained multiple authorities. You could see this on live on the multiple authorities search result page page title eg. "Error: Find a Disabled Students' Allowance assessment provider - GOV.UK"

This PR already adds the correct page title for search result pages, but the location error would still get appended to that eg. "Find a Disabled Students' Allowance assessment provider: select an address - GOV.UKError: Find a Disabled Students' Allowance assessment provider - GOV.UK"

The check added in this commit is the same as the check on line 17 to check if the multiple authorities template should be rendered (that line also checks if postcode is provided, but we already check for that on line 55 here).

However, I'm not sure if my "fix" is the best way to handle this or if bypassing multiple authorities in this way could be introducing some other issue. Maybe we should why the places_not_found method from the model doesn't recognise multiple authorities as a valid result. I've tried to superficially compare how this is handled for the other postcode finders like local transaction but those use the location_api model instead.