Turn CSRF protection on by default
The rails default CSRF protection was disabled in https://github.com/alphagov/frontend/pull/4123

The POST routes in this application are postcode lookups to an external API.
So we don't currently need CSRF protection, but this is being flagged by the scanner
as a vulnerability. Let's turn it back on across the application, and just disable it
on four specific routes. This means if the application changes in the future we aren't caught
out (assuming we're covered by default CSRF protection).