Get the base_path for the page requested.
This allows us to use the country specific path and slug
e.g. /foreign-travel-advice/france to get the content item, not just the
index.

Travel advice has parts so it's possible that the page requested is not
the base path

e.g. https://www.gov.uk/foreign-travel-advice/azerbaijan/safety-and-security
rather than https://www.gov.uk/foreign-travel-advice/azerbaijan

So we need to extract the base path to get the content item.

The default, request.path, should be moved to ContentItemController.