Add a PartsNavigationHelper
The presentation logic in the government-frontend presenter that wasn't
appropriate to include in the model.

Some methods have not been included as they are no longer being used in
government-frontend:

- parts_navigation
- parts_navigation_second_list_start
- part_links
- part_navigation_group_size

The render_guide_as_single_page? check has been removed from
part_link_elements as it is out of date. The election it refers to
was in 2021

See:
https://github.com/alphagov/government-frontend/blob/16a0e99c52f482f54875eaecf537b31e6e2b75b4/app/presenters/content_item_presenter.rb#L106

https://github.com/alphagov/government-frontend/blob/16a0e99c52f482f54875eaecf537b31e6e2b75b4/app/presenters/content_item_presenter.rb#L123-L126