Add a PartsNavigationHelper
Adds the presentation logic in the government-frontend presenter that it
wasn't appropriate to include in the TravelAdvice model.

Some methods have not been included as they are no longer being used in
government-frontend:

- parts_navigation
- parts_navigation_second_list_start
- part_links
- part_navigation_group_size

The render_guide_as_single_page? check has been removed from
part_link_elements as it is out of date. The election it refers to
was in 2021

Though this helper was taken from the travel advice code in
government-frontend, the helper name has been kept generic so that it
can be reused when guide pages are moved to frontend.

Audit trail:
content_item_presenter.rb:
1. https://github.com/alphagov/government-frontend/blob/16a0e99c52f482f54875eaecf537b31e6e2b75b4/app/presenters/content_item_presenter.rb#L106
2. https://github.com/alphagov/government-frontend/blob/16a0e99c52f482f54875eaecf537b31e6e2b75b4/app/presenters/content_item_presenter.rb#L123-L126