Add a Withdrawal concern.
Travel Advice pages need to known the withdrawn status in able to
construct the page title.

See:
https://github.com/alphagov/government-frontend/blob/main/app/presenters/content_item/withdrawable.rb#L8

and

https://github.com/alphagov/government-frontend/blob/main/app/presenters/travel_advice_presenter.rb#L6-L12

However, some of the methods feel like model concerns rather than
presentation concerns, so they have been moved the to model.

There is only one example of the content item with a withdrawn notice
in content schemas, so the shared example is being tested on the
generic content item model rather than travel advice.

The travel advice page title is a presentation concern and will be
added in a later commit.