Add a new presenter for content items
This presenter is badly named ContentItemModelPresenter because
ContentItemPresenter already exists. ContentItemPresenter takes a
hash of the results from content store and models them. That work is
already being done by the ContentItem model, and shouldn't be repeated
in the presenter.

Some extra work needs to be done to evaluate the existing routes to see
how they can be unstitched from the existing ContentItemPresenter and
then this class can be removed.

The page_title has been added to a generic presenter because other
document types from government-frontend use it as the basis for their
own page title methods.