Show metadata in the first part
This takes the metadata presenter method from government-frontend:
https://github.com/alphagov/government-frontend/blob/main/app/presenters/travel_advice_presenter.rb#L14-L28

but rather than recreating it, builds the params directly in the view.
The old presenter method need to use view_context to call the
`simple_format` helper method, which feels unnecessary for a method
that's only used once in this document type specific partial.

This code could probably be refactored into a more elegant helper method
in the future.