Add landing_page controller, routes, and basic view
A new flexible content type called a landing page is being created.
The content type will allow publishers to pick the "blocks" that
should be included on the page, e.g. govspeak, tabs, image, video,
and decide on the order of these "blocks".

As a content schema doesn't exist for the landing_page type yet,
a fake content item has been added to the controller. This will be
removed in a later PR once a content schema exists.

The additional content data path has been created as a constant to
allow it to be overwritten with fixtures in the tests, otherwise we'd
be testing with live content.