Add a block model
It is likely that the different types of blocks will need their own
models for their specific data requirements that builds on the base
block. A "factory" is being used to determine which model to use in a
similar way to that used to build the content item.

However general ancestors are used to check that the block model is of
the correct type rather than a direct ancestor as it will be necessary
to have blocks within blocks to be able to handled different types of
column blocks in later commits.

Co-authored-by: Keith Lawrence <Keith.Lawrence@digital.cabinet-office.gov.uk>