Use full width layout, contain blocks by default
Some blocks (in particular the hero image) should be full width.

The default layout puts a width-container on the entire content, making it difficult to do wider blocks without horrible CSS hacks.

Instead, we should use the full width layout, and put width containers only on the blocks that need them. The base class has this as false because full width blocks are the exception.

Note that unfortunately gem_layout_full_width is missing the blue bar, and possibly some other things we will want. We might need a new layout in the gem for this page.