Add hero block
A hero block has an image, and some content (which is a list of blocks).

The image should have 6 resolutions, for 3 viewports (mobile / tablet / desktop) and 2 pixel densities (default and 2x). These are rendered in a picture element, so we can do art direction as well as optimise the image size for individual users.

The hero content blocks are rendered in a loop, like they are in show.html. We pass inverse: true to them, as they're going to be light text on a dark background.

Note that the CSS is prototype-quality - we'll probably want to polish this a fair bit before go live.