Wrap all blocks in divs with conditional width class
Previously we were only wrapping blocks with divs if we needed to because we wanted to class.

It's probably generally useful to have them wrapped, and it makes the code a bit simpler.

I've added block-type specific classes, which are currently not used but might be useful for debugging / tweaking styles later.