Download requested chart
This is a bit ugly, but it works. I'm sure there's a much nicer way to
get this working.

Interesting tidbit is that even though the filename is passed in as
"chart_one.csv" what is actually returned by params[:filename] is
"chart_one". the ".csv" is dropped.

It's also a bit ugly that the download path is being defined directly
in the controller.

Renamed the chart_data folder to statistics. Should probably rename
everything to Statistics. That'll be the recommendation for the
production code.