Fix content store response type
The code was sometimes passing GdsApi::Responses into the ContentItem constructor, and sometimes Hashes.

GdsApi::Response behaves enough like Hash to be confusing, but it's not exactly the same. For example, it doesn't have a deep_merge method. It's also missing things which aren't to do with the body, like a cache_control method.

Technically the ContentItem constructor requires a GdsApi::Response, because it
delegates the cache_control method to that object. So we shouldn't have been passing it a Hash.

This commit adds a temporary `override_content_store_hash` parameter, which we can use to override the bits we want to without needing to pass the wrong type of object to the constructor. We'll be able to clean all this up once the content store returns the whole object.