Add a build_all helper method to BlockFactory
We're finding there are lots of places where we want to take a list of block hashes and turn them into a list of blocks (not all of these have hit main yet, but the need is there)

Currently we're duplicating the bit mapping code, and it's a bit ugly. This helper method DRYs things up a bit.