Add yaml content for landing page links
These links will power a side navigation block that will appear on some
landing pages.

The side navigation for all of the landing pages will have the same
links, so rather than adding a links list to the YAML for every landing
page, a shared one will be used.

This is done for expediency and does not reflect how the links will
eventually be handled by the CMS.

The side navigation have been nested inside a links folder to logically
separate them from the content items, and in case we need another list
of common links.