Add three column example of card blocks to landing page
The homepage design shows a three column area on desktop, each containing a
card element. This PR adds these elements to the yaml file so that card blocks
can be previewed in a 3 column context.

Thge columns use the previously created `grid-container` rather than the
standard row/columns because we need the images in the cards to line up
regardless of how much content precedes them in the card. This can only be
achieved using flexbox or grid when the cards are all in the same container.

It's not possible to align the images in the cards if each card remains in
a separate column.