Implement a blocks container block
This blocks container is used as an empty unstyled parent container to
hold other elements. Use it in situations where we don't want to create
a row/grid layout to contain blocks.

Consider this example:

```
- type: two_column_layout
  theme: one_third_two_thirds
  blocks:
  - type: govspeak
    content: <p>Left content!</p>
  - type: blocks_container
    blocks:
      - type: govspeak
        content: |
          <h2>This is a heading</h2>
          <p>Lorem ipsum...</p>
      - type: action_link
        text: "See the missions"
        href: "todo"

```

This example uses the `blocks_container` for the element type in the
second two-thirds width column. Without this empty unstyled container
we'd have no way of inserting multiple different blocks in the column.