Add landing page request test to increase test coverage
There are two ways that a landing page can be loaded. One is from a
content item, and one is by using the fake data in the controller.

This adds modifies the request tests to use the new example landing page
from content-schemas, and adds a test to check that the page
renders when the fake data in the controller is used.