Add a helper to render the contents list
It will be used to help render the side and main navigation blocks that
need to know which link to set to active.

This method is a helper as it needs to be called from the view to be
able to access `request.path` for the current path so that the link
can be set to "active".