Update main_navigation links hash in YAML files
This changes it to match the labels in the contents_list component,
as well as the forthcoming side_navigation block.

Removes links being hard-coded to active as this will be set my the
new helper.

Also updates the YAML sub-page-1 so that the active link can be
demonstrated.

Fix up to links hash