Update the button text on place finders
This an accessibility fix required by the DAC audit:
Update the button text from "Find" to "Find results near you"

For one place page, the Find a register office, the new button text
is "Find a register office".

The code uses the content of the publication title to determine that this is
the find a register service, rather than checking the base_path or content id, which would be
the usual way of doing it. This is intentional, because the button text only makes
sense on this page if the title matches. If in the future the title of this publication
changes, the button text will revert to the more generic "Find results near you". If we
were checking basepath and the page title changed, the button text would drift out of sync.