Render breadcrumbs in sub header
This keeps the breadcrumbs in the sub header with the dark background. Otherwise, the breadcrumbs render above the sub header on a white background. On a separate note, due to the current absence of a content item, I've used Hash.new as the content item to prevent the breadcrumb component from erroring. Additionally, due to the current absence of a content item, both sets of breadcrumbs will currently render; this will be resolved once we pass through the content item.