Add a render_block helper
The render_block method fails silently if a block doesn't exist.
This means that the page won't fail if a block is added to the YAML
before the block has been built, or if the block name has a typo in it.
The page won't fail, the block just won't display.