Add a rows method to return a row for each y-axis variable
Uses a `find` with an include to get the existing row. This is because
find does fuzzy match and returns all of the things it finds, where
as `include` does an exact match, but returns a boolean.

The CSV rows are being returned with symbol keys to that `to_sym`
doesn't need to be used in multiple places.

Finally, the row keys and values are being obfuscated.

We can't guarantee what the keys will be named so it's better not to
rely on them.

We're assuming that each row will have 3 columns:

- x-axis value
- variable / line name
- y-axis value

Co-authored-by: Andy Sellick <andy.sellick@digital.cabinet-office.gov.uk>