Allow the button text to match the publication title
Following a content review, it has been decided that on the
Find a register office page the button text should match the page title.

The code uses the content of the publication title to determine that this is
the find a register service, rather than checking the base_path or content id, which would be
the usual way of doing it. This is intentional, because the button text only makes
sense on this page if the title matches. If in the future the title of this publication
changes, the button text will revert to the more generic "Find results near you".