Add YAML content and data for a statistics block
Statistics data will be read in from a CSV to be provided by departments
and then converted into a chart in the statistics block on landing
pages.

It is assumed that users will not need to be able to download the CSV
files directly and that a link will be provided to an externally hosted
data source, e.g. ONS data.