Refactor js class name on menu button
As far as I can tell, the only thing that JS was doing with this class was removing it i.e. the class hook wasn't needed for some JS functionality, it was just removed when the JS initialised. As far as I'm aware, the purpose of this class is to hide the menu button when JS is unavailable/disabled, and it doesn't exist when JS is available/enabled, so I felt using a --no-js modifier was more suitable/easier to read.