Namespace SideNavigation block model under LandingPage
At the moment blocks are only used by Landing Page documents.
The block views are namespaced under LandingPage and the models should
be too to keep consistency, and so that it's clearer where blocks are
supposed to be used.

The stubbing of the constant to pick up the side navigation content has
also been fixed.