Update content_item model
- To expose first_public_at and first_published_at attributes
- Add Organisations concern
- Add contributors method which calls the organisations function from Organisations concern
to allow us to have a default value for all document_types

We need both first_public_at and first_published_at as these both need
to be checked otherwise updates won't be picked up for the other document types
that use this concern.

We have added a contributors function which returns the list of contributors,
these are then rendered in the "from" section of the publisher_metadata partial.
The contributors function will default to return just organisations for all document_types
however this function will be overriden depending on what each document_type needs.