Build statistics block from new data format
Previously we're expecting statistics data to be represented in three
columns, one of which would contain the variable name.

That is no longer the case. The variable name is coming from the Y-axis
header. Now the first column represents the x-axis and all of the
others represent the values on the y-axes.

The statistics block is still able to display multiple lines on a line
graph with the same x-axis.

The first header value is being ignored because that is the value for
the x-axis. The rows method is concerned with collating the values to
plot the data against the y-axis.

Also updates the tests to handle files where some data points are
missing, or provided as decimal values. And updates the landing page
factory to accept four attachments to test these different scenarios.

We no longer need to symbolize the keys in the CSV as we need to be able
to match against the header value which is a string, and pass the header
values for the y-axis to the chart component as a string.