Support draft CSVs
URI.open doesn't work on draft CSVs, because you need to provide
authorisation headers of some kind to prove you're allowed to access the
draft.

Instead, we need to use the asset manager API, as we do with CSV
previews.

Unfortunately Whitehall doesn't provide the ID we need to call asset
manager in a convenient form. Instead, we're parsing the URL and
extracting it from there. There's a TODO in the code because it would be
much nicer if whitehall provided the ID so we didn't have to do this
parsing.

I've been strict with the format of the ID (24 hexadecimal digits)
because I think it's better for the parsing to fail than to proceed with
a bad ID and fail later with a bad request. This does mean that if the
format of asset manager IDs changes this will break, but hopefully we
will have updates Whitehall to provide the ID directly long before that
happens.