Add govuk_chat_promo_helper
We want to show a promo to the new GOV.UK Chat application on certain pages
on GOV.UK.

This helper adds a list of base paths we want to render the Chat call to
action on.

If the GOVUK_CHAT_PROMO_ENABLED env var is set to true and the base
path passed in matches a base path from the constant the #show_govuk_chat_promo?
method returns true.