Add GOV.UK Chat entry banner to the base_page partial
This adds the chat entry banner to the base_page partial to the base_page
partial which is used for all content types.

This handles rendering the GOV.UK Chat Entry banner if the content items
base path is in the list of paths defined in the GovukChatPromoHelper.

The base paths we're adding only render this on select transation and smart
answer pages so i've only added the shared examples to the request spec for
those content types.

I can def see the argument for adding this to all content types that use the
base_page partial as there's nothing stopping someone adding base paths
for them and rendering the chat entry banner on ohter content types
pages.

Happy to discuss in the PR.