Make image blocks responsive
The block documentation incorrectly stated that image blocks took a list of sources in the same way that hero / featured blocks do.

This is the format that whitehall is going to send, so frontend needs to support it.

Since whitehall is providing lots of different image resolutions, we may as well provide markup so that the browser can work out what to do.

I believe this combination of srcset / sizes attributes on images in new to GOV.UK. I'm following the MDN guidance here:

https://developer.mozilla.org/en-US/docs/Learn/HTML/Multimedia_and_embedding/Responsive_images#resolution_switching_different_sizes

Translating this into this specific circumstance, what this HTML is saying is:

- We have six image sources, with intrinsic widths of 630, 1260, 708, 1416 etc.
- When the (max-width: 640px) breakpoint is true, we expect the element to be about 610px wide
- When the (max-width: 768px) breakpoint is true, we expect the element to be about 707px wide
- Otherwise, we expect the element to be about 630px wide

The browser will then use its knowledge of pixel density etc. to pick the best image source from the srcset.

This is a better approach than using the picture element when art-direction isn't a requirement, which for plain image blocks it isn't.