Make image blocks width: 100% by default
Otherwise the image size will just be the file size, which at small resolutions can push the content outside the edge of the viewport leading to a horizontal scrollbar.

We might want some other themes in future (e.g. two-thirds width images), but we haven't got a use case for those yet, and it feels like taking the whole width of the container is a good default.