Use large headings for default hero and featured blocks
This will cause all headings which are children of default hero or
featured blocks to get font_size = "l"

Previously this only happened to headings which were children of
heros with position: top, which isn't configured everywhere. Now it
happens for all heros with the default (non-middle_left) theme.

It's a bit of a breach of block isolation for hero / featured blocks to
be saying "if there's a child heading, it should be large". It might be
nicer just to provide the heading size in the YAML.

For now though, heading_font_size is a much clearer name than position,
and the heading block no longer has to work out "if position is top then
I should be large", which really felt like a lot more knowledge about
what its parents get up to at night than any block should have to have.

I've also moved a bit of configuration into the model so it's tested,
and a bit less verbose.