CSS for large govspeak in default hero images
We're overriding the rules set in markdown-typography in govuk_publishing_components for the tags there's a realistinc chance we might use.

I haven't overridden the styles for h1s or h4-6s, because they're not likely to be used in hero_content.

Note I've used govuk-font-size instead of govuk-font because it's only the size we're changing - we don't need to re-include the CSS for bold etc.