Pass font_size instead of position to headings
This way headings don't need to know about their parents, and whether
they're a top or not. Which lets face it, nobody wants to know.

Instead, the hero block knows about headings so it can pass down some
config to them.

We could do this with CSS, as we're doing for govspeak, but the
component isolation is better this way.