Add breadcrumb tests to the travel advice pages
The breadcrumb was initially missed off the pages, and then added
directly to the view, however as with the other pages in frontend the
application layout determines whether to set the layout.

These tests have been added to make sure that the breadcrumb continues
to display on the travel advice country pages, and does not display on
the index page.