Move breadcrumb rendering on travel advice pages
As with the other pages in frontend, the travel advice pages should be
able to rely on the application layout to render breadcrumbs.

For travel advice pages, breadcrumb will only be rendered if the
following condition is true:

`unless !publication`

This means that a `publication` object needs to exist. A conditional has
been added as we don't want the breadcrumb to display on the travel
advice index page.