Replace HomepagePresenter with a model
We currently have two ways to model content items and two content item
presenters.

* ContentItem models the response from content store
* ContentItemModelPresenter takes a ContentItem object and adds a
presentation layer to it.
* ContentItemPresenter does both, it models the response from content
store and adds presentation to it.

Ideally there would be content item models that inherit from ContentItem
and content item presenters that inherit from ContentItemModelPresenter.
And when the original ContentItemPresenter is removed, to rename
ContentItemModelPresenter to ContentItemPresenter.

To achieve this we need to look at the existing presenters one by one
and move the logic to the appropriate place.

In the homepage presenter all of the methods looked like they could be
moved to a model. The only exception was the call to a locale file.
Normally a call like this should be in a helper or presenter, however in
this case the locale is being used as data. Ideally data this should be
moved to a data file and a lookup to content store used to get the title
of the link. This will be fixed in a later change.