Remove hard-coded popular links
Now that popular links on the Homepage can be edited in a publisher
app and can be picked up from the content item, it's not clear whether
this hard-coded list of links is needed.

If they are, then this commit and the next one can be deleted. We might
want to consider picking these lists of links up from a "data" file
rather than a locale as strictly speaking they are not translated
content.

The test that checks the content of the popular links is being
displayed has been moved from the request tests to the system tests
so that it's with the other tests that check the page content.