Add slug to the SimpleSmartAnswer model
It is a copy of the method from ContentItemPresenter. The method will
not be removed from ContentItemPresenter until no routes are using it.

At the moment "slug" is only used by simple-smart-answers, so
eventhough it seems generic, it is not being added to the parent class
until another routes needs it.