Create a SimpleSmartAnswer model
We currently have two ways to model content items and two content item
presenters.

* ContentItem models the response from content store
* ContentItemModelPresenter takes a ContentItem object and adds a
presentation layer to it.
* ContentItemPresenter does both, it models the response from content
store and adds presentation to it.

Ideally there would be content item models that inherit from ContentItem
and content item presenters that inherit from ContentItemModelPresenter.
And when the original ContentItemPresenter is removed, to rename
ContentItemModelPresenter to ContentItemPresenter.

To achieve this we need to look at the existing presenters one by one
and move the logic to the appropriate place.

In the SimpleSmartAnswerPresenter only the "pass through" methods that
model items from the content store need to be moved to the model.
A `start_button_text` attribute as been added so that it can be used in
the `start_button_text` method in the presenter, the only method that
should remain.

The presenter will then be updated to accept an instance of this model
as a parameter in the next commit.