Show a breadcrumb on the travel advice index page
Follows on from: https://github.com/alphagov/frontend/pull/4511

It's not clear if this was purposefully left off, or was overlooked
because the travel advice controller didn't set the `@publication`
variable, however it is odd that the travel advice index page doesn't
have a breadcrumb. The other comparable page is the list of departments,
and that page does have a breadcrumb.